--- include/equation_answer.h
@@ -1,16 +1,0 @@
-/*
- * Copyright (c) Raphael Grubbauer
- * Licensed under the RGrubbauer Open Source License (GOSL) v1.4.0
- * See LICENSE.md file in the project root for full license information.
-*/
-
-#ifndef EQUATION_ANSWER_H
-#define EQUATION_ANSWER_H
-
-#include <iostream>
-#include <sstream>
-#include <string>
-
-float getEquationAnswer(std::string originalEquation);
-
-#endif
