--- include/generate_equation.h
@@ -1,16 +1,0 @@
-/*
- * Copyright (c) Raphael Grubbauer
- * Licensed under theGrubbauer Open Source License (GOSL) v1.4.0
- * See LICENSE.md file in the project root for full license information.
-*/
-
-#ifndef GENERATE_EQUATION_H
-#define GENERATE_EQUATION_H
-
-#include <string>
-
-#include "random.h"
-
-std::string randEquation(int lvl);
-
-#endif
