+++ include/grubbauer/savefile.h
@@ -4,15 +4,14 @@
  * See LICENSE.md file in the project root for full license information.
 */
 
-#ifndef OPEN_SAVEFILE_H
-#define OPEN_SAVEFILE_H
+#ifndef GRUBBAUER_SAVEFILE_H
+#define GRUBBAUER_SAVEFILE_H
 
-#include <filesystem>
-#include <fstream>
-#include <iostream>
-#include <nlohmann/json.hpp>
-#include <windows.h>
+#include <string>
 
-std::string openSaveFile();
+namespace grubbauer {
+  std::string readSaveFile();
+  void saveSaveFile(int lvl);
+}
 
 #endif
